Extract non-standard logging functions to different library
As part of the conversation in https://github.com/github/orchestrator/issues/472
this is a proof of concept to attempt to isolate non-standard logging
functions, so a more flexible and standard-lib compliant logging solution
can be dropped in.

This is accomplished by duplicating the non-standard functions from https://github.com/github/orchestrator/blob/master/vendor/github.com/openark/golib/log/log.go
into a new library.

It remains to be seen how these non-standard logging functions can have their
formatting remain consistent with different logging libraries, but for now,
this is simply an experiment to gauge support for this approach.