Allow dbdeployer/setup.sh to be reusable
Previously this script could only be used to setup a single instance.
I want to be able to reuse the script for the "frontend" database
setup for use in a local development environment. Therefore,
we need to move more of it's responsibilities (like actually calling `deploy single`)
into the integration test script, which seems reasonable.

It's to bad we have to grep for words out of the output of the script, but
it's the best I could think to do.